@@ -4,6 +4,6 @@ namespace Env0.Terminal.Config.Pocos
 {
     public class Filesystem
     {
-        public Dictionary<string, FileEntry> Root { get; set; } = new Dictionary<string, FileEntry>();
+        public Dictionary<string, FileEntry> Root { get; set; } = new();
     }
 }
